requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==4.1
django-cors-headers==4.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.7
PyJWT==2.8.0
pytz==2024.1
requests==2.32.3
sqlparse==0.5.0
typing-extensions==4.12.2
urllib3==2.2.2
